backend/requirements.txt
django>=4.0.0,<4.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
pyyaml==6.0
requests==2.28.2
django-cors-headers==3.13.0
Pillow==9.4.0
psycopg2==2.9.5
psycopg2-binary==2.9.5